requirements.txt
google-cloud-vision>=3.4.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
python-dotenv>=1.0.0
